+++ UnrealKit/UnrealKit.Core/Parsing/IStaticCameraHtmlReportService.cs
@@ -1,0 +1,8 @@
+﻿namespace UnrealKit.Core.Parsing;
+
+public interface IStaticCameraHtmlReportService
+{
+    Task<StaticCameraHtmlReportResult> GenerateAsync(
+        StaticCameraHtmlReportRequest request,
+        CancellationToken cancellationToken = default);
+}